update
did some significant change to the transaction file. I worked on the transaction record method. I've been able to write the code for the transaction record/receipt. I called the class at the end and commented it so you guys can uncomment and see the result. But its not fully done since we will still send these data to the database. And in turn retrieve the data from the database when needed and assign them to the right object.